Text.TeXMath.Readers.MathML.EntityMap

Description

HTML entity definitions as provided by W3C.

The mapping matches the version from 10th April 2014.

The original source can be downloaded from http://www.w3.org/TR/2014/REC-xml-entity-names-20140410/.

Note: I have made one alteration, switching epsilon and varepsilon, because the meanings of these names in HTML is different from the meanings in MathML+LaTeX. See http:/www.w3.org2003entities2007doc/#epsilon.

Documentation

getUnicode :: String -> Maybe String

Translates MathML entity reference to the corresponding Unicode string.