Text.TeXMath.Readers.MathML.MMLDict

Description

Dictionary of operators to MathML attributes as specified by the W3C standard.

The original file can be downloaded from here

Documentation

getMathMLOperator :: String -> FormType -> Maybe Operator

Tries to find the Operator record based on a given position. If there is no exact match then the positions will be tried in the following order (Infix, Postfix, Prefix) with the first match (if any) being returned.

operators :: [Operator]

A table of all operators as defined by the MathML operator dictionary.