Text.TeXMath.Readers.TeX.Macros

Description

Functions for parsing LaTeX macro definitions and applying macros to LateX expressions.

Documentation

data Macro

parseMacroDefinitions :: String -> ([Macro], String)

Parses a string for a list of macro definitions, optionally separated and ended by spaces and TeX comments. Returns the list of macros (which may be empty) and the unparsed portion of the input string.

applyMacros :: [Macro] -> String -> String

Applies a list of macros to a string recursively until a fixed point is reached. If there are several macros in the list with the same name, earlier ones will shadow later ones.