Text.TeXMath.Unicode.ToASCII

Description

This module exposes functions which attempt to approximate unicode characters as ASCII.

Information taken from the Unidecode python package which is based upon the Text::Unidecode Perl Module by Sean M. Burke.

Documentation

getASCII :: Char -> String

Approximates a single unicode character as an ASCII string (each character is between 0x00 and 0x7F).