Text.TeXMath.Unicode.ToTeX

Description

This module is derived from the list of unicode to LaTeX mappings compiled by Günter Milde.

An unmodified original copy of this work can be obtained from here

Documentation

getTeXMath :: String -> Env -> [TeX]

Converts a string of unicode characters into a strong of equivalent TeXMath commands. An environment is a list of strings specifying which additional packages are available.

getSymbolType :: Char -> TeXSymbolType

Returns TeX symbol type corresponding to a unicode character.

records :: [Record]